+++ app/Http/Controllers/TaskController.php
@@ -10,6 +10,7 @@ use App\Models\Client;
 use App\Models\Project;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -72,6 +73,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task): RedirectResponse
     {
+        Gate::authorize(PermissionEnum::DELETE_TASKS->value);
+
         $task->delete();
 
         return redirect()->route('tasks.index');
